node:fs: share path validation helpers via internal/validators
getValidatedPath and throwIfNullBytesInFileName are used by both the
lazily loaded internal/fs/watchfile module and the Windows realpath
fallbacks in node:fs, so export them from internal/validators instead of
keeping a copy in each file.
