+++ src/js/internal/validators.ts
@@ -96,6 +96,20 @@ function validateInternalField(object, fieldKey, className) {
   }
 }
 
+/** Validate a string-or-URL path and return it resolved to an absolute path string. */
+function getValidatedPath(p: any) {
+  if (p instanceof URL) return Bun.fileURLToPath(p as URL);
+  if (typeof p !== "string") throw $ERR_INVALID_ARG_TYPE("path", "string or URL", p);
+  if (p.startsWith("file:")) return Bun.fileURLToPath(p);
+  return require("node:path").resolve(p);
+}
+
+function throwIfNullBytesInFileName(filename: string) {
+  if (filename.indexOf("\u0000") !== -1) {
+    throw $ERR_INVALID_ARG_VALUE("path", "string without null bytes", filename);
+  }
+}
+
 hideFromStack(validateLinkHeaderValue, validateInternalField);
 hideFromStack(validateString, validateFunction, validateBoolean, validateUndefined);
 
@@ -143,4 +157,8 @@ export default {
   isUint8Array: value => value instanceof Uint8Array,
   /** `(object, fieldKey, className)` */
   validateInternalField,
+  /** `(path)` — accepts a string or file URL, returns it resolved to an absolute path string */
+  getValidatedPath,
+  /** `(filename)` */
+  throwIfNullBytesInFileName,
 };
